Updated namespaces
DCP became DistributedComputingPlatform and it is required so I added
code to make sure it is there with a nilREason if necessary.

Having a problem changing MD_ContentTypeCode to
MD_CoverageContentTypeCode. I created a template for this element, but
the * template is still being executed so I am getting two contentType
elements…