Renamed pngs
The image files that were created in the last pull requested had names
that were inconsistent with the html files. The standard names for
these images is xxx.png where xxx is the namespace prefix. If different
names are used, the links from the index.html files do not work. I
removed the old images and renamed the new ones.