A gmd:date with a nilReason attribute is emptied.
When a gmd:date has a nilReason attribute (with whatever value eg. withheld), the element is emptied.
The nilReason attribute should be copied but the children should not be affected by this attribute.
Also a gco:DateTime is xs:dateTime and can't have attribute. No need to copy the nilReason attribute to this element.