Fixed the class order in 2010-07-30
The schemas that were generated on 2010-07-30 had very different class
orders than the existing NGDC schemas. This made it difficult to fit
the new schemas into the git repository because most of the differences
were not important. This branch fixed the order of the classes so that
the more critical differences could be easily identified and examined.