Import Responsible Party templates
The gmd:CI_ResponsibilityParty template had to be included at the end
of the standard transform rather than imported at the beginning. This
ensures that the gmd:CI_ResponsibleParty template is used, but I’m not
sure why the import does not work. When I used the import, the
translations worked fine, but the root element remained
CI_ResponsibleParty rather than changing to CI_Responsibility.