Valid transform result
I have finally achieved a valid transform result with the 2014-12-25
scheme. Thanks to everyone that helped. The output is in
standards.iso.org/19115/-3/mdt/1.0/2014-12-25/mdt.xml. It is what i
call a PathFile that has the source 19115 paths as element contents.
This allow us to see where the content comes from in the source 19115
record.