Added iso directory to the hierarchy
Standards.iso.org includes an iso directory above all of the standard
directories so the correct path to 19115 is standards.iso.org/iso/19115
and to 19157 is standards.iso.org/iso/19115. That directory was
inadvertently left out of the git repository. Including it does not
effect the relative schema paths so it seems reasonable to introduce it
for conformance.