Restructure of ISO 19115-X
Disaggregation of ISO 19115-3 has resulted in flowon to several other standards.
To accomidate the upcoming release of ISO 19130-3 two versions have been generated V 2.1 to align with the 19130 document and 1.3 to incorporate the agreed numbering system. It is V 1.3 that will be used as the basis of future developments (it aligns with the CD for ISO 19115-3).